Fix so that multiplayer playing fields aren't duplicated
Sometimes when players drop out and join in again, playing fields are
duplicated.  I suspect that this is caused by the socket ID for a given
player changing.  Instead of relying on a persistent socket ID, we now
generate a UUID and identify players by that.

We only use the socket ID in the rare case when we haven't received the
UUID for a player yet.