Change so that view creation doesn't depend on model object instantiation order
This was often the cause of bugs or things not working for unintuitive
reasons.  Before, the views were created based on events in the model
object's constructors.  With this change, the views are created only
when those models are added to a higher-level construct.  You can now
instantiate model objects in any order.